Extract Stack class to manage Rule state during Recurrence iteration
