+++ lib/montrose/rule.rb
@@ -41,7 +41,6 @@ require "montrose/rule/hour_of_day"
 require "montrose/rule/month_of_year"
 require "montrose/rule/nth_day_of_month"
 require "montrose/rule/nth_day_of_year"
-require "montrose/rule/stack"
 require "montrose/rule/time_of_day"
 require "montrose/rule/total"
 require "montrose/rule/week_of_year"
